requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.15.0
xlwings==0.30.10
pandas==2.0.2
pyjwt[crypto]==2.6.0
